[datakit] Use shuffle for the clustered store
Replace the map-side store path with the shuffle implementation as the single canonical build_clustered_store API. The reference pipeline now configures reducers and subshards explicitly and hashes output-shaping values.

The 0.1B sample-mode pipeline completed end to end in 33m42s. The store emitted 20,162 records and 19,907,340 tokens through 64 reducers with zero job failures.
